+++ model_parameters.yml
@@ -1,0 +1,45 @@
+aggregator: add
+batch_size: 32
+cache_dir: /data/scratch/bjing/confgen/cache_alignHs/
+cache_path: /data/scratch/bjing/confgen/cache_alignHs//drugs_pos0_lim0_std
+data_dir: data/DRUGS/drugs/
+dataset: drugs
+fix_sigma: null
+in_edge_features: 4
+in_node_features: 74
+limit_train_mols: 0
+log_dir: workdir/alignHs30_l4_s20_2orderfix
+lr: 0.001
+max_radius: 5.0
+model: tensor_product
+n_epochs: 250
+no_batch_norm: false
+no_cache: false
+no_hydrogen: false
+no_residual: false
+ns: 32
+num_conv_layers: 4
+num_intermediate_conv_layers: 2
+num_workers: 1
+nv: 8
+optimizer: adam
+pos_enc_dim: 0
+radius_embed_dim: 50
+restart_dir: null
+run_name: alignHs30_l4_s20_2orderfix
+save_all_epochs: false
+scale_by_sigma: true
+scheduler: plateau
+scheduler_patience: 20
+seed: 0
+sigma_embed_dim: 32
+sigma_max: 3.14
+sigma_min: 0.031400000000000004
+split_path: data/DRUGS/splits/split0.npy
+std_pickle_size: 1000
+std_pickles: data/DRUGS/drugs_standardized_multi_assignment_alignHs/
+test_sigma_intervals: 0
+tp_aggregator: mean
+tp_force_aggregator: mean
+use_second_order_repr: true
+wandb: true
