+++ small_refined.smi
@@ -1,0 +1,25 @@
+[NH3+][C@@H](CCC(=O)N[C@@H](CSCc1ccccc1)C(=O)N[C@@H](C(=O)[O-])c1ccccc1)C(=O)[O-]	10gs.ligand
+CC(C)Cc1ccccc1	184l.ligand
+c1ccc2[nH]ccc2c1	185l.ligand
+CCCCc1ccccc1	186l.ligand
+Cc1ccc(C)cc1	187l.ligand
+Cc1ccccc1C	188l.ligand
+CCCC[C@@H]1CCCN(C(=O)[C@H](CCC(=O)[O-])NC(=O)[C@H](Cc2ccc(O[P+](=O)(=O)[O-])cc2)NC(C)=O)C1	1a1e.ligand
+CC(=O)[C@H]1CC[C@H]2[C@@H]3CCC4=CC(=O)CC[C@]4(C)[C@H]3CC[C@]12C	1a28.ligand
+CC(C)C[C@H](NC(=O)[C@H](CC(=O)[O-])NC(=O)[C@@H]([NH3+])CCC(=O)[O-])C(=O)[O-]	1a30.ligand
+CC(=O)Nc1ccc(N2C(=O)[C@@H]3[C@H]4CC[C@](NC(=O)OCC(=O)[O-])(CC4)[C@@H]3C2=O)cc1	1a4k.ligand
+Nc1nc2c(ncn2[C@@H]2O[C@H](CO[P@](=O)([O-])O[P@](N)(=O)[O-])[C@@H](O)[C@H]2O)c(=O)[nH]1	1a4r.ligand
+CN(C)c1cccc2c(S(=O)(=O)N[C@@H](CCCNC(N)=[NH2+])C(=O)N3CCCC[C@@H]3CCC(=O)c3nccs3)cccc12	1a4w.ligand
+O=c1[nH]cnc2c([C@@H]3O[C@H](CO)[C@@H](O)[C@H]3O)n[nH]c12	1a69.ligand
+CCCC[C@H](NC(=O)[C@H](C)NC(=O)[C@H](CCC(=O)[O-])NC(=O)[C@H](Cc1ccccc1)NC(=O)[C@H](CC(C)C)NC(=O)[C@@H](NC(=O)[C@@H]([NH3+])CCCNC(N)=[NH2+])C(C)C)C(N)=O	1a94.ligand
+[NH3+]CCCC[NH3+]	1a99.ligand
+CC(=O)N[C@@H](CC1CCCCC1)[C@@H](O)C[C@H](C(=O)N[C@@H](CCC(N)=O)C(=O)N[C@@H](CCCNC(N)=[NH2+])C(N)=O)C(C)C	1a9m.ligand
+O=c1[nH]cnc2[nH]cnc12	1a9q.ligand
+COC(=O)[C@@H](NC(=O)[C@@H](NC(=O)CC[C@H](O)[C@H](Cc1ccccc1)NC(=O)[C@H](C)NC(=O)[C@H](C)[NH3+])C(C)C)C(C)C	1aaq.ligand
+Nc1ccnc2c1ncn2[C@@H]1O[C@H](CO)[C@@H](O)[C@H]1O	1add.ligand
+CCCCC/C=C\C/C=C\C/C=C\C/C=C\CCCC(=O)[O-]	1adl.ligand
+O=C(CO)CO[P+](=O)(=O)[O-]	1ado.ligand
+Nc1ncnc2c1ncn2[C@@H]1O[C@H](CO[P@](=O)([O-])O[P+](=O)(=O)[O-])[C@@H](O[P+](=O)(=O)[O-])[C@H]1O	1afk.ligand
+Nc1ncnc2c1ncn2[C@@H]1O[C@H](CO[P@](=O)([O-])O[P+](=O)(=O)[O-])[C@@H](O)[C@H]1O[P+](=O)(=O)[O-]	1afl.ligand
+O=C([O-])Cc1ccc(O)c(O)c1	1ai4.ligand
+O=C([O-])Cc1cccc([N+](=O)[O-])c1	1ai5.ligand
