+++ small_refined_smiles_only.smi
@@ -1,0 +1,25 @@
+[NH3+][C@@H](CCC(=O)N[C@@H](CSCc1ccccc1)C(=O)N[C@@H](C(=O)[O-])c1ccccc1)C(=O)[O-]	
+CC(C)Cc1ccccc1	
+c1ccc2[nH]ccc2c1	
+CCCCc1ccccc1	
+Cc1ccc(C)cc1	
+Cc1ccccc1C	
+CCCC[C@@H]1CCCN(C(=O)[C@H](CCC(=O)[O-])NC(=O)[C@H](Cc2ccc(O[P+](=O)(=O)[O-])cc2)NC(C)=O)C1	
+CC(=O)[C@H]1CC[C@H]2[C@@H]3CCC4=CC(=O)CC[C@]4(C)[C@H]3CC[C@]12C	
+CC(C)C[C@H](NC(=O)[C@H](CC(=O)[O-])NC(=O)[C@@H]([NH3+])CCC(=O)[O-])C(=O)[O-]	
+CC(=O)Nc1ccc(N2C(=O)[C@@H]3[C@H]4CC[C@](NC(=O)OCC(=O)[O-])(CC4)[C@@H]3C2=O)cc1	
+Nc1nc2c(ncn2[C@@H]2O[C@H](CO[P@](=O)([O-])O[P@](N)(=O)[O-])[C@@H](O)[C@H]2O)c(=O)[nH]1	
+CN(C)c1cccc2c(S(=O)(=O)N[C@@H](CCCNC(N)=[NH2+])C(=O)N3CCCC[C@@H]3CCC(=O)c3nccs3)cccc12	
+O=c1[nH]cnc2c([C@@H]3O[C@H](CO)[C@@H](O)[C@H]3O)n[nH]c12	
+CCCC[C@H](NC(=O)[C@H](C)NC(=O)[C@H](CCC(=O)[O-])NC(=O)[C@H](Cc1ccccc1)NC(=O)[C@H](CC(C)C)NC(=O)[C@@H](NC(=O)[C@@H]([NH3+])CCCNC(N)=[NH2+])C(C)C)C(N)=O	
+[NH3+]CCCC[NH3+]	
+CC(=O)N[C@@H](CC1CCCCC1)[C@@H](O)C[C@H](C(=O)N[C@@H](CCC(N)=O)C(=O)N[C@@H](CCCNC(N)=[NH2+])C(N)=O)C(C)C	
+O=c1[nH]cnc2[nH]cnc12
+COC(=O)[C@@H](NC(=O)[C@@H](NC(=O)CC[C@H](O)[C@H](Cc1ccccc1)NC(=O)[C@H](C)NC(=O)[C@H](C)[NH3+])C(C)C)C(C)C
+Nc1ccnc2c1ncn2[C@@H]1O[C@H](CO)[C@@H](O)[C@H]1O
+CCCCC/C=C\C/C=C\C/C=C\C/C=C\CCCC(=O)[O-]	
+O=C(CO)CO[P+](=O)(=O)[O-]	
+Nc1ncnc2c1ncn2[C@@H]1O[C@H](CO[P@](=O)([O-])O[P+](=O)(=O)[O-])[C@@H](O[P+](=O)(=O)[O-])[C@H]1O
+Nc1ncnc2c1ncn2[C@@H]1O[C@H](CO[P@](=O)([O-])O[P+](=O)(=O)[O-])[C@@H](O)[C@H]1O[P+](=O)(=O)[O-]	
+O=C([O-])Cc1ccc(O)c(O)c1	
+O=C([O-])Cc1cccc([N+](=O)[O-])c1
